bumper, Move cnap repo related tests to cnaoRepoCommands_test.go
We purely* move the unit-tests related to cnaoRepo to the dedicated
test file cnaoRepoCommands_test.go

* only exception is a small fix in isComponentBumpNeeded where we also
update the url so that the test won't fail.
